Implemented new collision system
A far simpler one, based on Unity's physics, rather than ugly hacks.
The motivation for this was a hard-to-replicate bug which caused ships
to sometimes fail to exit the "in collision" state, getting them stuck.
It was tested both on shielded and unshielded ships. Looks like this
solved the bug.