Implemented follow formation
At least on a basic level.
Ships can now be ordered to follow a ship, an they will form a chain. If a ship in the chain is disabled, it is "removed" from the chain.
Currently this behavior is still a bit prone to chaos and friendly fire.
Also moved more constants to the global constants file.